requirements.txt
fastapi
uvicorn
kafka-python
testcontainers
redis
joblib
scikit-learn
pandas
numpy
